- added more service functions - hooked up the object relations (job, payType, payCalendar, and payCalendarDates) - modified a couple field types
git-svn-id: https://svn.kuali.org/repos/kpme/trunk@301 6c5f4195-72ea-6250-e7e4-e498f552a5df

+++ src/main/java/org/kuali/hr/time/clock/location/ClockLocationRule.java
@@ -21,7 +21,7 @@ public class ClockLocationRule extends PersistableBusinessObjectBase {
 
 	private Long workArea;
 	private String principalId;
-	private BigInteger jobNumber;
+	private Long jobNumber;
 
 	private Date effectiveDate;
 	private boolean active;
@@ -97,11 +97,11 @@ public class ClockLocationRule extends PersistableBusinessObjectBase {
 		this.userPrincipalId = userPrincipalId;
 	}
 
-	public BigInteger getJobNumber() {
+	public Long getJobNumber() {
 		return jobNumber;
 	}
 
-	public void setJobNumber(BigInteger jobNumber) {
+	public void setJobNumber(Long jobNumber) {
 		this.jobNumber = jobNumber;
 	}
